patch to light effects and state handling
In this patch I added a GlobalHandler that can hold values of low
importance that don't need their own script such as if some random blood
splatters in a hidden room are visible or not. We decided to add this
for little things such as the lights being bright at the start of the
game but dimming at some point.

Co-authored by: @Dey3000